initial move gaz data provider out of fuzzy search and map components into own proivder
