+++ libraries/commons/utils/src/index.ts
@@ -2,6 +2,8 @@ export { suppressReactCismapErrors } from "./lib/log-react-cismap-errors";
 
 export { type CarmaConfig, extractCarmaConfig } from "./lib/carmaConfig";
 
+export { getGazData, type GazDataItem } from "./lib/gazData";
+
 export { generateRandomString } from "./lib/strings";
 
 export { getApplicationVersion } from "./lib/version";
